refactor(web): vitest設定をパッケージルートの1ファイルに集約
client/server/storybookの3設定とヘルパ(generate, workers-pool)を
packages/web/vitest.config.ts に統合。--mode で設定を出し分ける。

各targetはカバレッジのprovider(server=istanbul, client/storybook=v8系)・
閾値・実行環境が異なり1回のcoverage runに統合できないため、test.projects
ではなく--modeで分岐し、従来通り3つの独立したvitest runとして実行する。
Storybookビルダー経由(mode=development/production)ではdefaultで
storybook設定を返すためviteConfigPathも同一ファイルで解決できる。

package.json/.storybook/main.ts/CIのtest_covアクションの参照パスを更新。
