+++ application/packages/web/package.json
@@ -6,9 +6,9 @@
   "scripts": {
     "dev": "react-router dev",
     "build": "react-router build",
-    "test:client": "vitest run -c src/test/vitest/client/config.ts",
-    "test:server": "vitest run -c src/test/vitest/server/config.ts",
-    "test-storybook": "vitest run -c src/test/vitest/storybook/config.ts",
+    "test:client": "vitest run --mode client",
+    "test:server": "vitest run --mode server",
+    "test-storybook": "vitest run --mode storybook",
     "e2e": "playwright test --config=src/test/e2e/playwright.config.ts",
     "e2e:report": "playwright show-report src/test/e2e/playwright-report",
     "e2e:gen": "playwright codegen --config=src/test/e2e/playwright.config.ts",
