--- application/packages/web/src/test/vitest/workers-pool.ts
@@ -1,37 +1,0 @@
-import { dirname, resolve } from 'node:path'
-import { fileURLToPath } from 'node:url'
-import { cloudflarePool, cloudflareTest, readD1Migrations } from '@cloudflare/vitest-pool-workers'
-import type { Plugin } from 'vite'
-
-// migrations は datastore パッケージで一元管理しているため、パッケージ(packages/web)から相対参照する。
-const MIGRATIONS_DIR = resolve(
-  dirname(fileURLToPath(import.meta.url)),
-  '..',
-  '..',
-  '..',
-  '..',
-  'datastore',
-  'migrations',
-)
-
-export async function createWorkersPool(): Promise<{
-  plugins: Plugin[]
-  pool: ReturnType<typeof cloudflarePool>
-}> {
-  const migrations = await readD1Migrations(MIGRATIONS_DIR)
-
-  const poolOptions = {
-    miniflare: {
-      compatibilityDate: '2025-04-01',
-      compatibilityFlags: ['nodejs_compat'],
-      d1Databases: { DB: 'test-db' },
-      bindings: {
-        TEST_MIGRATIONS: migrations,
-        NODE_ENV: 'test',
-        LOG_LEVEL: 'silent',
-      },
-    },
-  }
-
-  return { plugins: [cloudflareTest(poolOptions)], pool: cloudflarePool(poolOptions) }
-}
